backend/requirements.txt
fastapi
pydantic
requests
openai
python-dotenv
uvicorn
pymupdf
rapidfuzz
pandas
python-multipart
tiktoken 
googlesearch-python 
newspaper4k 
nltk 
transformers 
lxml_html_clean
MarkItDown
docxtpl